feat: insertion and deletion for data.sakura

+++ evaluation/data/sakura.py
@@ -14,11 +14,21 @@ class SakuraDataProvider(RedisDataProvider):
             filter_type=RedisDataProvider.skip_type_enum_rev["skip_missing_without_warn"],
         )
 
-    def insert_ds(self):
-        ds = [{**d, "key": f"{self.data_id}:{i}"} for i, d in enumerate(get_sakura_ds())]
+    def insert_ds(self, is_exp=False):
+        if input(f"[SakuraDataProvider] are you sure to insert sakura? (is_exp={is_exp}) | (y/n) > ") != "y":
+            print("[SakuraDataProvider] insertion cancelled")
+            return
+
+        ds = [
+            {**d, "key": f"{self.data_id}:{i}", "status": "initialized"}
+            for i, d in enumerate(get_sakura_ds(is_exp=is_exp))
+        ]
         self.insert_samples(ds)
 
     def delete_ds(self):
+        if input(f"[SakuraDataProvider] are you sure to delete sakura? | (y/n) > ") != "y":
+            print("[SakuraDataProvider] deletion cancelled")
+            return
         self.delete_samples(self.get_all_keys())
 
 
@@ -26,7 +36,7 @@ if __name__ == "__main__":
     sakura_provider = SakuraDataProvider(
         redis_cfg=RedisConfig(host="salmonn.hufs.jae.one", port=6379, db=9),
         key_prefix="salmonn-13b:sakura",
-        filter={"set": "language"},
+        filter={"set": "language", "hop": "single"},
     )
 
     sakura_provider.delete_ds()
@@ -40,3 +50,4 @@ if __name__ == "__main__":
 
     print(f"data_id: {sakura_provider.data_id}")
     print(f"len(samples): {len(sakura_provider)}")
+    sakura_provider.status(keys=["status", "bin_inference"])
